Added the Navigation bar to React
