+++ app/globals.css
@@ -1,0 +1,9 @@
+:root {
+	--accent3: #8f1fffff;
+	--accent2: #d32289ff;
+	--accent1: rgb(21, 0, 213);
+	--secondary: rgb(23, 0, 88);
+	--primary: #0f071dff;
+	--border-color: #838383;
+	--text-color: #dfdfdf;
+}
